42. Strands Agent Usecase/Device Management Strands/requirements.txt

pandas>=2.0,<3.0
fastapi>=0.100,<1.0
uvicorn[standard]>=0.23,<1.0
pydantic>=2.0,<3.0
strands-agents[ollama]
strands-agents-tools
